backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyyaml
tldextract
python-docx
openpyxl
docxcompose
Pillow
